Завершение тестового задания, но не завершение работы

+++ app/controllers/BlogController.php
@@ -14,21 +14,23 @@ class BlogController extends Controller
     /**
      * @var Blog
      */
-    private $model;
+    protected $model;
     /**
      * @var BlogView
      */
-    private $view;
+    protected $view;
     /**
      * @var CommentController $comments
      */
-    private $comments;
+    protected $comments;
     /**
      * @var UserController $user
      */
-    private $user;
+    protected $user;
 
     protected $subTitle = "Блог";
+
+    protected $p;
 
 
     public function __construct()
@@ -44,41 +46,55 @@ class BlogController extends Controller
     public function actionIndex($routes)
     {
         $_SESSION['blog-post-id'] = false;
-
+        $p = filter_input(INPUT_GET, "page", FILTER_SANITIZE_NUMBER_INT);
+        $this->p = !$p ? 1 : $p;
         $data = array(
             'site-title' => $this->title,
             'page-header' => 'Блог'
         );
         $pageContent = "";
-        $content = $this->model->getData('blog', "", "LIMIT 0,5");
+        $limit = sprintf(" LIMIT %d, 5",
+            ($this->p - 1) * 5
+        );
+        $content = $this->model->getData('blog', "", $limit);
         foreach ($content as $id => $entry) {
-            $params = array(
-                'item-link' => '/blog/' . $entry['id'],
-                'item-title' => $entry['title'],
-                'item-cut-content' => $entry['cutcontent'],
-                'item-created' => $this->formatCreatedDate($entry['dateadd']),
-                'item-author' => $this->user->getName($entry['user_id']),
-                'item-views' => $this->pluralViews($entry['views']),
-                'item-comments' => $this->pluralComments($this->comments->getCountForPost($entry['id'])),
-                'item-poster' => $entry['poster'],
-            );
-            $pageContent .= $this->view->buildPartial("list-item", $params);
+            $pageContent .= $this->getListParams($entry);
         }
+        $pageContent .= $this->getPaginator('/blog/?');
+
         $data['page-content'] = $pageContent;
 
         $this->view->buildLayout(false, $data);
     }
 
+    protected function getListParams($entry)
+    {
+        $params = array(
+            'item-link' => '/blog/' . $entry['id'],
+            'item-title' => $entry['title'],
+            'item-cut-content' => $entry['cutcontent'],
+            'item-created' => $this->formatCreatedDate($entry['dateadd']),
+            'item-author' => $this->user->getName($entry['user_id']),
+            'item-views' => $this->pluralViews($entry['views']),
+            'item-comments' => $this->pluralComments($this->comments->getCountForPost($entry['id'])),
+            'item-poster' => $entry['poster'],
+        );
+        return $this->view->buildPartial("list-item", $params);
+    }
+
     /**
      * @param array|null $routes
+     * @param null $error
      */
-    public function actionShow($routes)
+    public function actionShow($routes, $error = null)
     {
-        $pageContent = "";
         $id = array_shift($routes);
+        $this->increaseCounter($id);
+        $pageContent = "";
         $_SESSION['blog-post-id'] = $id;
         $post = $this->model->getItem('blog', $id);
         $params = array(
+            'user-id' => $_SESSION['user_id'],
             'post-id' => $post['id'],
             'post-poster' => $post['poster'],
             'post-title' => $post['title'],
@@ -90,6 +106,10 @@ class BlogController extends Controller
             'post-created' => $this->formatCreatedDate($post['dateadd']),
             'post-comment-list' => $this->getCommentList($post['id'])
         );
+        if($error !== null){
+            $params['error'] = true;
+            $params['error-text'] = $error;
+        }
         $pageContent .= $this->view->buildPartial("item", $params);
         $this->title .= " - " . $post['title'];
         $data = array(
@@ -124,6 +144,10 @@ class BlogController extends Controller
             $data = $this->getFormData();
             if ($this->checkRequired($data, $params)) {
                 $postId = $this->model->insertData("blog", $data);
+
+                if (!$postId) {
+                    $this->buildError($params, $data, "Ошибка добавления материала");
+                }
                 $posterName = $this->loadPoster($postId);
                 if ($posterName != false) {
                     $this->model->update("blog", $postId, array(
@@ -132,6 +156,8 @@ class BlogController extends Controller
                 }
                 header("Location: /blog/" . $postId);
                 die();
+            } else {
+                $this->buildError($params, $data, "Заполнены не все поля отмеченные звездочкой '*'");
             }
         }
 
@@ -142,8 +168,7 @@ class BlogController extends Controller
      */
     public function actionEdit($routes)
     {
-        $postId = array_shift($routes);
-
+        $postId = intval(array_shift($routes));
         $params = array(
             'site-title' => $this->title . " - Изменение материала",
             'page-header' => 'Изменение материала',
@@ -152,7 +177,7 @@ class BlogController extends Controller
         );
         if ((User::logged() && BlogController::isAuthor()) || UserController::isAdminNow()) {
             $act = filter_input(INPUT_POST, "act", FILTER_SANITIZE_STRING);
-            if ($act != "edit") {
+            if ($act != "update") {
                 $post = $this->model->getItem("blog", $postId);
                 $params = array_merge(array(
                     'post-id' => $post['id'],
@@ -163,10 +188,31 @@ class BlogController extends Controller
                 ), $params);
                 $params['page-content'] = $this->view->buildPartial("form", $params);
                 $this->view->buildLayout(false, $params);
+                die();
             } else {
                 $data = $this->getFormData();
+
                 if ($this->checkRequired($data, $params)) {
-                    $this->model->update("blog", $postId, $data);
+                    $result = $this->model->update("blog", $postId, $data);
+                    if (!$result) {
+                        $this->buildError($params, $data, "Ошибка изменения материала");
+                    }
+                    $posterName = $this->loadPoster($postId);
+                    if ($posterName != false) {
+                        $this->model->update("blog", $postId, array(
+                            "poster" => $posterName
+                        ));
+                    } else if (filter_input(INPUT_POST, "delete-poster", FILTER_SANITIZE_NUMBER_INT) == 1) {
+                        {
+                            $this->deletePoster($postId);
+                            $this->model->update("blog", $postId, array(
+                                "poster" => false
+                            ));
+                        }
+
+                    }
+                    header("Location: /blog/" . $postId);
+                    die();
                 }
             }
         } else {
@@ -180,12 +226,33 @@ class BlogController extends Controller
      */
     public function actionDelete($routes)
     {
-        $postId = array_shift($routes);
-        if ($this->model->delete("blog", $postId)) {
-            header("Location: /blog/");
+        $postId = intval(array_shift($routes));
+        if ((User::logged() && BlogController::isAuthor()) || UserController::isAdminNow()) {
+            if ($this->model->delete("blog", $postId)) {
+                $this->deletePoster($postId);
+                header("Location: /blog/");
+                die();
+            } else {
+                $this->actionShow([$postId], "Ошибка при удалении страницы");
+            }
         } else {
-            header("Location: /blog/" . $postId);
+            header("Location: /blog/");
+            die();
         }
+    }
+
+    protected function buildError($params, $data, $errorString)
+    {
+        $params = array_merge(array(
+            'post-title' => $data['title'],
+            'post-cutcontent' => $data['cutcontent'],
+            'post-content' => $data['content'],
+            'error' => true,
+            'error-text' => $errorString
+        ), $params);
+        $params['page-content'] = $this->view->buildPartial("form", $params);
+        $this->view->buildLayout(false, $params);
+        die();
     }
 
     /**
@@ -209,7 +276,7 @@ class BlogController extends Controller
      * @param string $date
      * @return false|string
      */
-    private function formatCreatedDate($date)
+    protected function formatCreatedDate($date)
     {
         return date('d.m.Y в H:i', strtotime($date));
     }
@@ -218,7 +285,7 @@ class BlogController extends Controller
      * @param int $n
      * @return string
      */
-    private function pluralComments($n)
+    protected function pluralComments($n)
     {
         if (!$n) {
             return "Комментариев нет";
@@ -231,7 +298,7 @@ class BlogController extends Controller
      * @param int $n
      * @return string
      */
-    private function pluralViews($n)
+    protected function pluralViews($n)
     {
         if (!$n) {
             return "Просмотров нет";
@@ -244,7 +311,7 @@ class BlogController extends Controller
      * @param integer $postId
      * @return false|integer
      */
-    private function getAuthorId($postId)
+    protected function getAuthorId($postId)
     {
         return $this->model->getAuthor($postId);
     }
@@ -271,13 +338,17 @@ class BlogController extends Controller
      */
     protected function getFormData()
     {
-        return array(
+        $data = array(
             "title" => filter_input(INPUT_POST, "post-title", FILTER_SANITIZE_STRING),
             "cutcontent" => htmlspecialchars_decode(filter_input(INPUT_POST, "post-cutcontent", FILTER_SANITIZE_SPECIAL_CHARS)),
             "content" => htmlspecialchars_decode(filter_input(INPUT_POST, "post-content", FILTER_SANITIZE_SPECIAL_CHARS)),
-            "user_id" => $_SESSION['user_id'],
-            "dateadd" => date("Y-m-d H:i:s"),
         );
+        if ($_POST['act'] == "insert") {
+            $data["poster"] = "";
+            $data["user_id"] = $_SESSION['user_id'];
+            $data["dateadd"] = date("Y-m-d H:i:s");
+        }
+        return $data;
     }
 
     /**
@@ -293,11 +364,47 @@ class BlogController extends Controller
                 'post-cutcontent' => $data['cutcontent'],
                 'post-content' => $data['content'],
             ), $params);
-            $params['page-content'] = $this->view->buildPartial("form", $params);
-            $this->view->buildLayout(false, $params);
-            die();
+            return false;
         }
         return true;
+    }
+
+    /**
+     * @param int $postId
+     */
+    private function deletePoster($postId)
+    {
+        $post = $this->model->getItem('blog', $postId);
+        $posterPath = sprintf("%s/img/%s", $_SERVER['DOCUMENT_ROOT'], $post['poster']);
+        unlink($posterPath);
+    }
+
+    /**
+     * @param int $postId
+     */
+
+    private function increaseCounter($postId)
+    {
+        if (!User::logged() || !BlogController::visited($postId)) {
+            $post = $this->model->getItem("blog", $postId);
+            $viewCounter = $post['views'] + 1;
+            $this->model->update("blog", $postId, array("views" => $viewCounter));
+            $_SESSION['visited'][$postId] = 1;
+        }
+    }
+
+    protected function getPaginator($link, $where = '')
+    {
+
+        $itemCount = $this->model->getCount('blog', $where);
+
+
+        $data = array(
+            "page-list-count" => floor($itemCount / 5),
+            'page-list-current' => $this->p,
+            'page-list-link' => $link
+        );
+        return $this->view->buildPartial('page-list', $data);
     }
 
     /**
@@ -316,5 +423,11 @@ class BlogController extends Controller
             return true;
         }
         return false;
+    }
+
+    public static function visited($postId)
+    {
+        return $_SESSION['visited'][$postId];
+
     }
 }
